Added automatic claw servo sync with flip (with Neev)
